+++ packages/cli/src/__tests__/utils/constants.ts
@@ -1,4 +1,5 @@
-import { DEFAULT_EXPORT_DIRECTORY } from "../../commands";
+import { DEFAULT_EXPORT_DIRECTORY } from '../../commands';
 
 export const TEST_STORAGE_LOCATION = DEFAULT_EXPORT_DIRECTORY;
+export const TEST_ACCOUNT = 'account';
 export const TEST_API_KEY = 'apiKey';