+++ packages/integration-sdk-cli/src/commands/sync.ts
@@ -3,8 +3,9 @@ import { createCommand } from 'commander';
 import * as log from '../log';
 import {
   getApiKeyFromEnvironment,
+  getAccountFromEnvironment,
   getApiBaseUrl,
-  createApiClientWithApiKey,
+  createApiClient,
   createIntegrationLogger,
   synchronizeCollectedData,
 } from '@jupiterone/integration-sdk-runtime';
@@ -18,14 +19,18 @@ export function sync() {
     )
     .action(async (options) => {
       log.debug('Loading API Key from JUPITERONE_API_KEY environment variable');
-      const apiKey = getApiKeyFromEnvironment();
-      const apiBaseUrl = getApiBaseUrl({ dev: !!process.env.JUPITERONE_DEV });
+      const accessToken = getApiKeyFromEnvironment();
 
+      log.debug('Loading account from JUPITERONE_ACCOUNT environment variable');
+      const account = getAccountFromEnvironment();
+
+      const apiBaseUrl = getApiBaseUrl({ dev: !!process.env.JUPITERONE_DEV });
       log.debug(`Configuring client to access "${apiBaseUrl}"`);
 
-      const apiClient = createApiClientWithApiKey({
+      const apiClient = createApiClient({
         apiBaseUrl,
-        apiKey,
+        account,
+        accessToken,
       });
 
       const { integrationInstanceId } = options;