+++ packages/integration-sdk-runtime/src/api/error.ts
@@ -1,29 +1,19 @@
 import { IntegrationError } from '@jupiterone/integration-sdk-core';
 
-export class IntegrationMalformedApiKeyError extends IntegrationError {
-  constructor(message: string) {
-    super({
-      code: 'MALFORMED_API_KEY_ERROR',
-      message,
-    });
-  }
-}
-
 export class IntegrationApiKeyRequiredError extends IntegrationError {
-  constructor(message: string) {
+  constructor() {
     super({
-      code: 'JUPITER_ONE_API_KEY_ENVIRONMENT_VARIABLE_NOT_SET',
-      message,
+      code: 'JUPITERONE_API_KEY_ENVIRONMENT_VARIABLE_NOT_SET',
+      message: 'The JUPITERONE_API_KEY environment variable must be set',
     });
   }
 }
 
-export function malformedApiKeyError() {
-  return new IntegrationMalformedApiKeyError('Malformed API Key provided');
-}
-
-export function apiKeyRequiredError() {
-  return new IntegrationMalformedApiKeyError(
-    'The JUPITERONE_API_KEY environment variable must be set',
-  );
+export class IntegrationAccountRequiredError extends IntegrationError {
+  constructor() {
+    super({
+      code: 'JUPITERONE_ACCOUNT_ENVIRONMENT_VARIABLE_NOT_SET',
+      message: 'The JUPITERONE_ACCOUNT environment variable must be set',
+    });
+  }
 }