+++ packages/integration-sdk/CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to
 - Removed `createApiClientWithApiKey` helper from runtime package. You must use
   `createApiClient` for compatibility with new service tokens.
 
+### Fixed
+
+- Pass `onFailure` function to children of `IntegrationLogger`
+
 ## 3.8.0 - 2020-10-26
 
 ### Added